+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Breaking changes
 * @iov/bcp-types: Pluralize `.amounts` property and change to `Amount` array in
   `SwapOfferTransaction`, `SwapCounterTransaction`, `SwapData`
 * @iov/bcp-types: Rename `BcpAtomicSwap` to `AtomicSwap`
+* @iov/bcp-types: Rename atomic swap query types to `AtomicSwapQuery`,
+  `AtomicSwapRecipientQuery`, `AtomicSwapSenderQuery`, `AtomicSwapIdQuery`,
+  `AtomicSwapHashlockQuery`, `isAtomicSwap*Query`.
 
 ## 0.12.1